:sparkles: Creating 'reports' app
The idea is to create a 'reports_app' that will handle any report that a user want to do. Can be a report about an bad Attraction, a mean Comment, etc. Basically any kind of previously app we created.
All reports will be called by a custom action inside the ViewSet of the app that wants to implement this feature.