:children_crossing: Nested objects are read_only now
I plan to alter this. But for now, when consulting resources, the user cannot send a 'complete tourist spot object' add/edit at once, they need to send each part (object) at once.
It will be really easy to use in the front-end when changed, will be really similar to a no-realational database.